fix: ajuste en el nombre de la carpeta para los dete de ubicacion
