+++ EventResourceReservationApp.Application/DTOs/Locations/UpdateLocationRequest.cs
@@ -5,10 +5,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace EventResourceReservationApp.Application.DTOs.Loctions
+namespace EventResourceReservationApp.Application.DTOs.Locations
 {
-    public class CreateLocationRequest
+    public class UpdateLocationRequest
     {
+        [Required(ErrorMessage = "El Id de la ubicacion es obligatorio.")]
+        public int Id { get; set; }
         [Required(ErrorMessage = "El nombre del pais es obligatorio.")]
         [StringLength(100, ErrorMessage = "El nombre del pais no puede exceder los 100 caracteres.")]
         public string Country { get; set; }
@@ -33,8 +35,5 @@ namespace EventResourceReservationApp.Application.DTOs.Loctions
 
         [StringLength(50, ErrorMessage = "El número interior no puede exceder los 50 caracteres.")]
         public string InteriorNumber { get; set; }
-
-        [Required(ErrorMessage = "El ID del usuario creador es obligatorio.")]
-        public Guid CreatedByUserId { get; set; }
     }
 }
