+++ EventResourceReservationApp.Application/UseCases/Locations/ReadByIdLocationUseCase.cs
@@ -1,5 +1,5 @@
 ﻿using EventResourceReservationApp.Application.Common;
-using EventResourceReservationApp.Application.DTOs.Loctions;
+using EventResourceReservationApp.Application.DTOs.Locations;
 using EventResourceReservationApp.Application.Repositories;
 using EventResourceReservationApp.Domain;
 using Microsoft.Extensions.Logging;
@@ -45,7 +45,7 @@ namespace EventResourceReservationApp.Application.UseCases.Locations
                     ExteriorNumber = location.ExteriorNumber,
                     InteriorNumber = location.InteriorNumber,
                     CreatedByUserId = location.CreatedByUserId,
-                    CreateAt = location.CreatedAt
+                    CreatedAt = location.CreatedAt
                 };
                 return OperationResult<LocationResponse>.Success(response, "Ubicación encontrada exitosamente.");
             }
